pyproject.toml
[project]
name = "hack-agent-repo-it"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic",
    "crewai",
    "crewai-tools",
    "openai",
    "qdrant-client",
    "gitpython",
    "neo4j",
    "pinecone",
    "weaviate-client",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "python-dotenv",
    "httpx",
    "aiofiles",
    "jinja2",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "redis",
    "celery",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-community",
    "tree-sitter",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "tree-sitter-typescript",
    "tree-sitter-java",
    "tree-sitter-cpp",
    "tree-sitter-go",
    "tree-sitter-rust",
    "pygments",
    "markdown",
    "beautifulsoup4",
    "requests",
    "aiohttp",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "plotly",
    "streamlit",
]
